Changed to TypeScript
