+++ admin/tab_m.html
@@ -31,7 +31,7 @@
           }
       </style>
       <title>Device manager</title>
-    <script type="module" crossorigin src="./assets/app-995L7_Zr.js"></script>
+    <script type="module" crossorigin src="./assets/app-L1UZS96K.js"></script>
   </head>
   <body>
     <div id="root"></div>
