+++ src-admin/package.json
@@ -14,6 +14,8 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "@iobroker/adapter-react-v5": "^4.7.13",
+    "@iobroker/dm-utils": "^0.1.3",
+    "@iobroker/dm-gui-components": "^0.0.4",
     "@types/react": "^18.2.43",
     "@types/react-dom": "^18.2.17",
     "@vitejs/plugin-react": "^4.2.1",
