+++ src-admin/src/components/InstanceManager/index.js
@@ -1,0 +1,3 @@
+import DeviceList from './DeviceList';
+
+export default DeviceList;
