+++ src-admin/src/components/Page.tsx
@@ -1,0 +1,86 @@
+import {
+    Grid,
+    Paper,
+} from '@mui/material';
+
+import TopBar from './TopBar';
+// import DeviceList from '@iobroker/dm-gui-components';
+import DeviceList from './InstanceManager';
+import { getTranslation } from '@iobroker/dm-gui-components/Utils';
+import Communication, {CommunicationProps, CommunicationState} from './InstanceManager/Communication';
+import React from "react";
+
+interface PageProps extends CommunicationProps {
+    /* Instance to upload images to, like `adapterName.X` */
+    uploadImagesToInstance?: string;
+ }
+
+ interface PageState extends CommunicationState {
+     filter: string;
+ }
+
+/**
+ * Page - Main page
+ * @param {object} params - Parameters
+ * @param {object} params.socket - Socket object
+ * @param {object} params.uploadImagesToInstance - Instance to upload images to
+ * @returns {Element}
+ * @constructor
+ */
+export default class Page extends Communication<PageProps, PageState> {
+    constructor(props) {
+        super(props);
+        Object.assign(this.state, {
+            selectedInstance: window.localStorage.getItem('dm.selectedInstance') || '',
+            filter: '',
+        });
+    }
+
+    loadData(): Promise<void> {
+        return Promise.resolve(undefined);
+    }
+
+    renderContent() {
+        /** @type {object} */
+        const gridStyle = {
+            justifyContent: 'center',
+            alignItems: 'stretch',
+        };
+        /** @type {object} */
+        const emptyStyle = {
+            padding: 25,
+        };
+
+        // TODO: Show toast while action cannot be executed because of broken session or reload page
+
+        return <div className="App-header" style={{ width: '100%', height: '100%', overflow: 'hidden' }}>
+            <Paper elevation={1} style={{ width: '100%', height: '100%', overflow: 'hidden' }}>
+                <TopBar
+                    selectedInstance={this.state.selectedInstance}
+                    setSelectedInstance={selectedInstance => {
+                        window.localStorage.setItem('dm.selectedInstance', selectedInstance);
+                        this.setState({ selectedInstance });
+                    }}
+                    filter={this.state.filter}
+                    setFilter={filter => this.setState({ filter })}
+                    socket={this.props.socket}
+                    instanceHandler={this.instanceHandler}
+                />
+                <div style={{ width: '100%', height: 'calc(100% - 66px)', overflow: 'auto', marginTop: 8 }}>
+                    <Grid container style={gridStyle}>
+                        {!this.state.selectedInstance && <div style={emptyStyle}>
+                            <span>{getTranslation('noInstanceSelectedText')}</span>
+                        </div>}
+                        <DeviceList
+                            embedded
+                            uploadImagesToInstance={this.props.uploadImagesToInstance}
+                            selectedInstance={this.state.selectedInstance}
+                            filter={this.state.filter}
+                            socket={this.props.socket}
+                        />
+                    </Grid>
+                </div>
+            </Paper>
+        </div>;
+    }
+}
